--- myplayer/src/App.css
@@ -1,51 +1,0 @@
-.container {
-  text-align: center;
-}
-
-video {
-  max-width: 100%;
-  height: auto;
-}
-
-.playIcon {
-  display: block;
-  width: 50px;
-  position: relative;
-  top: calc(-50px - 10vh);
-  margin: 0 auto;
-  background-color: white;
-  border-radius: 50%;
-}
-
-@media (min-width: 400px) {
-  .playIcon {
-    top: calc(-71px - 10vh);
-  }
-}
-
-@media (min-width: 530px) {
-  .playIcon {
-    top: calc(-120px - 10vh);
-  }
-}
-
-@media (min-width: 650px) {
-  .playIcon {
-    top: calc(-160px - 10vh);
-  }
-}
-
-@media (min-width: 800px) {
-  .playIcon {
-    top: calc(-215px - 10vh);
-    width: 100px;
-  }
-}
-
-
-@media (min-width: 1000px) {
-  .playIcon {
-    top: calc(-300px - 10vh);
-    width: 100px;
-  }
-}
